Change system to support library feature variants
