+++ dsl/common/src/main/groovy/net/neoforged/gradle/dsl/common/extensions/JarJar.groovy
@@ -1,51 +1,12 @@
 package net.neoforged.gradle.dsl.common.extensions
 
-import net.neoforged.gradle.dsl.common.dependency.DependencyFilter
-import net.neoforged.gradle.dsl.common.dependency.DependencyVersionInformationHandler
 import org.gradle.api.Action
 import org.gradle.api.artifacts.Dependency
-import org.gradle.api.publish.maven.MavenPublication
+import org.gradle.api.plugins.JavaPluginExtension
 
-interface JarJar {
+interface JarJar extends JarJarFeature {
 
     String EXTENSION_NAME = "jarJar";
-
-    /**
-     * Enable the jarJar default configuration, unless already disabled
-     */
-    void enable();
-
-    /**
-     * Disable the jarJar default configuration
-     */
-    void disable();
-
-    /**
-     *
-     * @param disable or un-disable the jarJar default configuration; allows reversing {@link #disable()}.
-     */
-    void disable(boolean disable);
-
-    /**
-     * {@return whether the jarJar task should by default copy the contents and manifest of the jar task}
-     */
-    boolean getDefaultSourcesDisabled();
-
-    /**
-     * Stop the jarJar task from copying the contents and manifest of the jar task
-     */
-    void disableDefaultSources();
-
-    /**
-     * Set whether the jarJar task should copy the contents and manifest of the jar task
-     * @param value whether to disable the default sources
-     */
-    void disableDefaultSources(boolean value);
-
-    /**
-     * Set the jarJar task to source dependencies to include from the runtime classpath
-     */
-    void fromRuntimeConfiguration();
 
     /**
      * Sets the version of a dependency contained within a jarJar jar.
@@ -70,13 +31,10 @@ interface JarJar {
     void ranged(Dependency dependency, String range);
 
     /**
-     * Filter the dependencies included by the jarJar task
+     * Configure jarJar for a library feature with the given name, as created by {@link JavaPluginExtension#registerFeature(String,Action)}.
+     * Creates a featureJarJar task and configuration for the feature if they are missing.
+     * @param featureName the name of the feature to configure
+     * @return the configuration for jarJar for the feature
      */
-    JarJar dependencies(Action<DependencyFilter> c);
-
-    /**
-     * Configure the versions used by the dependencies included by the jarJar task
-     */
-    JarJar versionInformation(Action<DependencyVersionInformationHandler> c);
-
+    JarJarFeature forFeature(String featureName);
 }
